+++ Domain/Game/User.cs
@@ -4,7 +4,7 @@ namespace ChatGame.BL.Domain
 {
     public class User
     {
-        public ushort Id { get; set; }
+        public uint Id { get; set; }
         [Required]
         [RegularExpression("^[a-z_0-9]{4,25}$", ErrorMessageResourceName = "InvalidUserName", ErrorMessageResourceType = typeof(Resources.Resources))]
         public string UserName { get; set; }
